moved command into a related file
