--- frontend/src/ts/commandline/lists/copy-words-to-clipboard.ts
@@ -1,39 +1,0 @@
-import * as Misc from "../../utils/misc";
-import * as Notifications from "../../elements/notifications";
-
-const subgroup: MonkeyTypes.CommandsSubgroup = {
-  title: "Are you sure...",
-  list: [
-    {
-      id: "copyNo",
-      display: "Nevermind",
-    },
-    {
-      id: "copyYes",
-      display: "Yes, I am sure",
-      exec: (): void => {
-        const words = Misc.getWords();
-
-        navigator.clipboard.writeText(words).then(
-          () => {
-            Notifications.add("Copied to clipboard", 1);
-          },
-          () => {
-            Notifications.add("Failed to copy!", -1);
-          }
-        );
-      },
-    },
-  ],
-};
-
-const commands: MonkeyTypes.Command[] = [
-  {
-    id: "copyWordsToClipboard",
-    display: "Copy words to clipboard",
-    icon: "fa-copy",
-    subgroup,
-  },
-];
-
-export default commands;
