+++ frontend/src/ts/commandline/lists/result-screen.ts
@@ -1,6 +1,32 @@
 import * as TestLogic from "../../test/test-logic";
 import * as TestUI from "../../test/test-ui";
 
+const copyWords: MonkeyTypes.CommandsSubgroup = {
+  title: "Are you sure...",
+  list: [
+    {
+      id: "copyNo",
+      display: "Nevermind",
+    },
+    {
+      id: "copyYes",
+      display: "Yes, I am sure",
+      exec: (): void => {
+        const words = Misc.getWords();
+
+        navigator.clipboard.writeText(words).then(
+          () => {
+            Notifications.add("Copied to clipboard", 1);
+          },
+          () => {
+            Notifications.add("Failed to copy!", -1);
+          }
+        );
+      },
+    },
+  ],
+};
+
 const commands: MonkeyTypes.Command[] = [
   {
     id: "nextTest",
@@ -51,6 +77,12 @@ const commands: MonkeyTypes.Command[] = [
     available: (): boolean => {
       return TestUI.resultVisible;
     },
+  },
+  {
+    id: "copyWordsToClipboard",
+    display: "Copy words to clipboard",
+    icon: "fa-copy",
+    subgroup: copyWords,
   },
 ];
 export default commands;
